fix: write-failure guard returned success; pqueue priorities 32-bit on Windows
Two defects found in a post-merge review of the cleanup sweep, both in
code that sweep introduced.

close_generated_file's caller returned 1 when a generated stream failed
to write. compile_source reports failure as 0 and every caller tests
!compile_source(...), so the guard printed its error and then handed the
truncated .c to the C compiler anyway, which is the exact behaviour it
was added to prevent. It now returns 0.

std.pqueue's C entry points took `long` for the priority. Aether `long`
is 64-bit everywhere, but C `long` is 32-bit on Windows, so the extern
was ABI-mismatched there: priorities past 2^31 truncated, and the
regression test only passed because small values survive in the argument
register. The C side now uses `long long`, the same convention
string_to_long_raw documents for exactly this reason. The Aether-facing
module API is unchanged.

+++ std/collections/aether_pqueue.h
@@ -14,8 +14,10 @@ typedef struct AetherPQueue AetherPQueue;
 // Returns NULL on allocation failure.
 AetherPQueue* aether_pqueue_new(void);
 
-// 1 on success, 0 on a null queue or allocation failure.
-int aether_pqueue_push(AetherPQueue* pq, long priority, void* item);
+// 1 on success, 0 on a null queue or allocation failure. Priority is
+// long long, not long: Aether `long` is 64-bit and Windows C `long`
+// is 32, matching the string_to_long_raw convention.
+int aether_pqueue_push(AetherPQueue* pq, long long priority, void* item);
 
 // Removes and returns the lowest-priority item, or NULL when empty.
 // A NULL return is ambiguous if NULL items were pushed; use
@@ -27,7 +29,7 @@ void* aether_pqueue_peek(AetherPQueue* pq);
 
 // Priority of the item aether_pqueue_peek would return. Returns 0 when empty,
 // so check aether_pqueue_size first.
-long aether_pqueue_peek_priority(AetherPQueue* pq);
+long long aether_pqueue_peek_priority(AetherPQueue* pq);
 
 int aether_pqueue_size(AetherPQueue* pq);
 int aether_pqueue_is_empty(AetherPQueue* pq);
